backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==1.10.13
python-multipart==0.0.6
